requirements.txt
streamlit 
langchain-openai 
langchain-groq 
langchain-google-genai 
langchain-core
groq
sentence-transformers
faiss-cpu
PyPDF2
tavily-python
python-dotenv
langchain
langchain-community
